extracted safe submitter and event parser from cassandra output
