+++ lib/logstash/outputs/cassandra_output.rb
@@ -1,11 +1,11 @@
 # encoding: utf-8
 require "logstash/outputs/base"
 require "logstash/namespace"
-require "time"
 require "logstash/outputs/cassandra/buffer"
-require "cassandra"
+require "logstash/outputs/cassandra/event_parser"
+require "logstash/outputs/cassandra/safe_submitter"
 
-class LogStash::Outputs::ElasticSearch < LogStash::Outputs::Base
+class LogStash::Outputs::CassandraOutput < LogStash::Outputs::Base
 
     milestone 1
 
@@ -95,20 +95,20 @@ class LogStash::Outputs::ElasticSearch < LogStash::Outputs::Base
     config :idle_flush_time, :validate => :number, :default => 1
 
     def register()
-      @statement_cache = {}
       assert_filter_transform_structure(@filter_transform) if @filter_transform
+      setup_event_parser()
+      setup_safe_submitter()
       setup_buffer_and_handler()
-      assert_filter_transform_structure()
     end
 
     def receive(event)
-      @buffer << build_cassandra_action(event)
+      @buffer << @event_parser.parse(event)
     end
 
     # Receive an array of events and immediately attempt to index them (no buffering)
     def multi_receive(events)
       events.each_slice(@flush_size) do |slice|
-        safe_submit(slice.map {|e| build_cassandra_action(e) })
+        @safe_submitter.submit(slice.map {|e| @event_parser.parse(e) })
       end
     end
 
@@ -130,192 +130,21 @@ class LogStash::Outputs::ElasticSearch < LogStash::Outputs::Base
     end
 
     def setup_buffer_and_handler
-      @buffer = ::LogStash::Outputs::Buffer.new(@logger, @flush_size, @idle_flush_time) do |actions|
-        safe_submit(actions)
+      @buffer = ::LogStash::Outputs::CassandraOutput::Buffer.new(@logger, @flush_size, @idle_flush_time) do |actions|
+        @safe_submitter.submit(actions)
       end
     end
 
-    def setup_cassandra_session()
-      cluster = ::Cassandra.cluster(
-        username: @username,
-        password: @password,
-        hosts: @hosts,
-        consistency: @consistency.to_sym,
-        timeout: @request_timeout,
-        retry_policy: get_retry_policy(@retry_policy),
-        logger: @logger
+    def setup_safe_submitter()
+      @safe_submitter = ::LogStash::Outputs::Cassandra::SafeSubmitter.new(
+        @
       )
-      @session = cluster.connect(@keyspace)
-      @logger.info("New cassandra session created",
-                   :username => @username, :hosts => @hosts, :keyspace => @keyspace)
     end
 
-    def get_retry_policy(policy_name)
-      case policy_name
-        when "default"
-          return ::Cassandra::Retry::Policies::Default.new
-        when "downgrading_consistency"
-          return ::Cassandra::Retry::Policies::DowngradingConsistency.new
-        when "failthrough"
-          return ::Cassandra::Retry::Policies::Fallthrough.new
-      end
-    end
-
-    def build_cassandra_action(event)
-      action = {}
-      action["table"] = event.sprintf(@table)
-      filter_transform = get_filter_transform(event)
-      if filter_transform
-        action["data"] = {}
-        for filter in filter_transform
-          add_event_value_from_filter_to_action(event, filter, action)
-        end
-      else
-        add_event_data_using_configured_hints(event, action)
-      end
-
-      return action
-    end
-
-    def get_filter_transform(event)
-      filter_transform = nil
-      if @filter_transform_event_key
-        filter_transform = event.sprintf(@filter_transform_event_key)
-        assert_filter_transform_structure(filter_transform)
-      elsif @filter_transform
-        filter_transform = @filter_transform
-      end
-      return filter_transform
-    end
-
-    def add_event_value_from_filter_to_action(event, filter, action)
-      event_data = event.sprintf(filter["event_data"])
-      if filter.has_key?("cassandra_type")
-        cassandra_type = event.sprintf(filter["cassandra_type"])
-        event_data = convert_value_to_cassandra_type(event_data, cassandra_type)
-      end
-      column_name = event.sprintf(filter["column_name"])
-      action["data"][column_name] = event_data
-    end
-
-    def safe_submit(actions)
-      begin
-        batch = prepare_batch(actions)
-        @session.execute(batch)
-        @logger.info("Batch sent successfully")
-      rescue Exception => e
-        @logger.warn("Failed to send batch (error: #{e.to_s}).")
-      end
-    end
-
-    def prepare_batch(actions)
-      statement_and_values = []
-      for action in actions
-        query = "INSERT INTO #{@keyspace}.#{action["table"]} (#{action["data"].keys.join(', ')})
-          VALUES (#{("?" * action["data"].keys.count).split(//) * ", "})"
-
-        if !@statement_cache.has_key?(query)
-          @statement_cache[query] = @session.prepare(query)
-        end
-        statement_and_values << [@statement_cache[query], action["data"].values]
-      end
-
-      batch = @session.batch do |b|
-        statement_and_values.each do |v|
-          b.add(v[0], v[1])
-        end
-      end
-      return batch
-    end
-
-    def convert_value_to_cassandra_type(event_data, cassandra_type)
-      generator = get_cassandra_type_generator(cassandra_type)
-      typed_event_data = nil
-      begin
-        typed_event_data = generator.new(event_data)
-      rescue Exception => e
-        error_message = "Cannot convert `value (`#{event_data}`) to `#{cassandra_type}` type"
-        if @ignore_bad_values
-          case event_data
-            when 'int', 'varint', 'bigint', 'double', 'decimal', 'counter'
-              typed_event_data = 0
-            when 'uuid', 'timeuuid'
-              typed_event_data = generator.new("00000000-0000-0000-0000-000000000000")
-            when 'timestamp'
-              typed_event_data = generator.new(Time::parse("1970-01-01 00:00:00"))
-            when 'inet'
-              typed_event_data = generator.new("0.0.0.0")
-            when 'float'
-              typed_event_data = generator.new(0)
-            when 'boolean'
-              typed_event_data = generator.new(false)
-            when 'text', 'varchar', 'ascii'
-              typed_event_data = generator.new(0)
-            when 'blob'
-              typed_event_data = generator.new(nil)
-            when /^set\((.*)\)$/
-              typed_event_data = generator.new([])
-          end
-          @logger.warn(error_message, :exception => e, :backtrace => e.backtrace)
-        else
-          @logger.error(error_message, :exception => e, :backtrace => e.backtrace)
-          raise error_message
-        end
-      end
-      return typed_event_data
-    end
-
-    def get_cassandra_type_generator(name)
-      case name
-        when 'uuid'
-          return Cassandra::Types::Uuid
-        when 'timestamp'
-          return Cassandra::Types::Timestamp
-        when 'inet'
-          return Cassandra::Types::Inet
-        when 'float'
-          return Cassandra::Types::Float
-        when 'varchar'
-          return Cassandra::Types::Varchar
-        when 'text'
-          return Cassandra::Types::Text
-        when 'blob'
-          return Cassandra::Types::Blog
-        when 'ascii'
-          return Cassandra::Types::Ascii
-        when 'bigint'
-          return Cassandra::Types::Bigint
-        when 'counter'
-          return Cassandra::Types::Counter
-        when 'int'
-          return Cassandra::Types::Int
-        when 'varint'
-          return Cassandra::Types::Varint
-        when 'boolean'
-          return Cassandra::Types::Boolean
-        when 'decimal'
-          return Cassandra::Types::Decimal
-        when 'double'
-          return Cassandra::Types::Double
-        when 'timeuuid'
-          return Cassandra::Types::Timeuuid
-        when /^set\((.*)\)$/
-          set_type = get_cassandra_type_generator($1)
-          return Cassandra::Types::Set(set_type)
-        else
-          raise "Unknown cassandra_type #{cassandra_type}"
-      end
-    end
-
-    def add_event_data_using_configured_hints(event, action)
-      action["data"] = event.to_hash()
-      # Filter out @timestamp, @version, etc to be able to use elasticsearch input plugin directly
-      action["data"].reject!{|key| %r{^@} =~ key}
-      @hints.each do |event_key, cassandra_type|
-        if action["data"].has_key?(event_key)
-          event_data = convert_value_to_cassandra_type(action["data"][event_key], cassandra_type)
-        end
-      end
+    def setup_event_parser()
+      @event_parser = ::LogStash::Outputs::Cassandra::EventParser.new(
+        @table, @filter_transform_event_key, @filter_transform, @hints, @ignore_bad_values, @logger
+      )
     end
   end
 end end end
